requirements.txt
streamlit
librosa
numpy
pandas
scikit-learn
pydub
joblib
streamlit-audiorecorder
